Too much too really keep track of. Kinda Works now

+++ WordleConsole/ConsoleApp1/Classes/WordleSession.cs
@@ -3,27 +3,45 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Text.Json;
+using System.Text.Json.Serialization;
 using System.Threading.Tasks;
+using System.IO;
+using System.Security.AccessControl;
+using System.Security.Principal;
+
 
 namespace Wordle.Classes
 {
-    
+
     public class WordleSession
     {
+        [JsonPropertyName("PlayerName")]
         public string PlayerName { get; set; }
-        public Wordlist Dict { get; set; }
-        public Wordlist Fails { get; set; }
-        public Wordlist Solved { get; set; }
-        public int Difficulty { get; set; }
+        [JsonPropertyName("Dict")]
 
-        public WordleSession (string name)
+        public Wordlist Dict { get; set; }
+        [JsonPropertyName("Fails")]
+        public Wordlist Fails { get; set; }
+        [JsonPropertyName("Solved")]
+        public Wordlist Solved { get; set; }
+        [JsonPropertyName("Difficulty")]
+        public int Difficulty { get; set; }
+        [JsonPropertyName("Score")]
+        public int Score { get; set; }
+
+
+        //Constructor for deserialization
+
+        public WordleSession() { }
+        public WordleSession(string name)
         {
             PlayerName = name;
-            Dict = new Wordlist ();
+            Dict = new Wordlist(1);
             Dict.ReadListFromFile();
-            Fails = new Wordlist ();
-            Solved = new Wordlist ();
+            Fails = new Wordlist(1);
+            Solved = new Wordlist(1);
             Difficulty = 5;
+            Score = 0;
         }
         public void SetDifficulty()
 
@@ -33,7 +51,7 @@ namespace Wordle.Classes
             Console.WriteLine("Please Type in how many tries you want to guess a word.\nStandard difficulty is 5, you can choose a value between 3-7");
             while (value < 3 || value > 7)
             {
-               
+
                 value = inget.GetIntFromKey();
                 if (value < 3 || value > 7)
                 {
@@ -46,23 +64,25 @@ namespace Wordle.Classes
 
         public void StartTry()
         {
+            bool success = false;
             int counter = 0;
             WordlePresenter presenter = new WordlePresenter();
             WordleLogic logic = new WordleLogic(this);
             logic.PickWordFromValidSelection();
             while (!logic.MatchesWord() && counter < Difficulty)
             {
-                
+
                 presenter.AskForGuess();
                 logic.AskForInput();
                 logic.CompareTryToWord();
                 counter++;
                 Console.Clear();
                 presenter.DisplayAllGuesses(logic);
-                
+
             }
             if (logic.MatchesWord())
             {
+                success = true;
                 presenter.ShowSuccess(logic);
                 Solved.Words.Add(logic.CurrentWord);
                 Dict.Words.Remove(logic.CurrentWord);
@@ -73,29 +93,50 @@ namespace Wordle.Classes
                 presenter.ShowFailure(logic);
                 Fails.Words.Add(logic.CurrentWord);
                 Dict.Words.Remove(logic.CurrentWord);
-                presenter.AnyKey();
+
             }
+
+            Console.WriteLine($"\nYou scored {ScorePoints(logic.CurrentWord, counter, Difficulty, success)} points");
+            UpdateHighscore();
+            presenter.AnyKey();
+
         }
-        
+
         public void SaveGame()
         {
-            Console.WriteLine("Please type the Path where you want your savefile located:\nFor example: c:\\wordle\\save\\mysave");
+            Console.WriteLine("Please type the Path where you want your savefile located:\nFor example: c:\\wordle\\save\\mysave\nMake sure you can write into the directory");
 
             string path = Console.ReadLine();
-            string jsonsession = JsonSerializer.Serialize<WordleSession>(this);
+            string directoryPath = Path.GetDirectoryName(path);
+            JsonSerializerOptions options = new JsonSerializerOptions { WriteIndented = true };
+            string jsonsession = JsonSerializer.Serialize<WordleSession>(this, options);
+            if (Directory.Exists(directoryPath))
 
-            try
             {
-                StreamWriter writer = new StreamWriter(path, false);
+                try
+                {
 
-                writer.Write(jsonsession);
+                    StreamWriter writer = new StreamWriter(path, false);
 
-                writer.Close();
+                    writer.Write(jsonsession);
+
+                    writer.Close();
+                }
+                catch (Exception ex)
+                {
+                    Console.WriteLine(ex.Message);
+                }
             }
-            catch (Exception ex)
+            else
             {
-                Console.WriteLine(ex.Message);
+                Console.Clear();
+                Console.WriteLine("\tSavegame can not be written, because Directory doesn't exists.");
+                Console.WriteLine("\n\npress any key to go back to main menu");
+                Console.ReadKey(intercept: true);
+                Console.Clear();
             }
+
+
         }
 
         public void ChooseDict()
@@ -108,19 +149,19 @@ namespace Wordle.Classes
             {
                 case 1:
                     this.Dict.Path = "data\\words.txt";
-                    this.Fails = new Wordlist();
-                    this.Solved = new Wordlist();
+                    this.Fails = new Wordlist(1);
+                    this.Solved = new Wordlist(1);
                     this.Dict.ReadListFromFile();
                     break;
                 case 2:
                     this.Dict.Path = "data\\words2.txt";
-                    this.Fails = new Wordlist();
-                    this.Solved = new Wordlist();
+                    this.Fails = new Wordlist(1);
+                    this.Solved = new Wordlist(1);
                     this.Dict.ReadListFromFile();
                     break;
                 case 3:
                     break;
-                    default:
+                default:
                     break;
             }
 
@@ -133,5 +174,79 @@ namespace Wordle.Classes
             presenter.Statistics(this);
         }
 
+        public double ScorePoints(string currentword, int tries, int difficulty, bool success)
+        {
+            double pointgain = 0;
+            if (!success)
+            {
+                pointgain = -5;
+            }
+            else
+            {
+                pointgain = Math.Ceiling((10 + (currentword.Length - tries + 1 * 2)) * (1 + 0.2 * (difficulty - 5)));
+            }
+            Score += (int)pointgain;
+            return pointgain;
+        }
+
+        public void UpdateHighscore()
+        {
+            HighscoreEntry info = new HighscoreEntry { Playername = this.PlayerName, Score = this.Score };
+            string path = "data\\highscore";
+
+
+            HighScore hs = new HighScore { HighscoreList = new List<HighscoreEntry>() };
+
+            try
+            {
+
+                if (File.Exists(path))
+                {
+                    string json = File.ReadAllText(path);
+                    hs = JsonSerializer.Deserialize<HighScore>(json);
+                }
+            }
+            catch (Exception ex)
+            {
+                Console.WriteLine($"Error reading highscore file: {ex.Message}");
+            }
+
+            bool updated = false;
+
+
+            for (int i = 0; i < hs.HighscoreList.Count; i++)
+            {
+                if (hs.HighscoreList[i].Playername == PlayerName)
+                {
+                    if (info.Score > hs.HighscoreList[i].Score)
+                    {
+                        hs.HighscoreList[i] = info;
+                    }
+                    updated = true;
+                    break;
+                }
+            }
+
+
+            if (!updated)
+            {
+                hs.HighscoreList.Add(info);
+            }
+
+
+            hs.HighscoreList.Sort((x, y) => y.Score.CompareTo(x.Score));
+
+            try
+            {
+
+                string json = JsonSerializer.Serialize(hs, new JsonSerializerOptions { WriteIndented = true });
+                File.WriteAllText(path, json);
+            }
+            catch (Exception ex)
+            {
+                Console.WriteLine($"Error writing highscore file: {ex.Message}");
+            }
+        }
     }
-}
+
+    }
